Organize code to extension
